Convert backend to Go
Removes the NodeJS backend server and replaces it with a Go
implementation.
Functionality on the React side is pretty much the same. Some small
tweaks to the return data from the API calls needed to be made for now.

Also changed the forecast component/API to just return the current
weather instead of a 3 day spread.